refactor: Moved leave components from components/leaves to features/leaves/components
