+++ app/users/dashboard/workforce/leaves/approvals/page.tsx
@@ -20,9 +20,9 @@ import {
   SelectValue,
 } from '@/components/ui/select';
 import { Pagination, SearchAndFilter } from '@/components/common';
-import { StatCard } from '@/components/leave/stat-card';
-import { TableSkeleton } from '@/components/leave/skeletons';
-import { EmptyState } from '@/components/leave/empty-state';
+import { StatCard } from '@/features/leave/components/stat-card';
+import { TableSkeleton } from '@/features/leave/components/skeletons';
+import { EmptyState } from '@/features/leave/components/empty-state';
 import { useApproverRequests } from '@/hooks/leave/use-leave';
 import {
   LeaveStatus,
