+++ app/users/dashboard/workforce/leaves/requests/page.tsx
@@ -20,10 +20,10 @@ import {
   SelectValue,
 } from '@/components/ui/select';
 import { Pagination, SearchAndFilter } from '@/components/common';
-import { LeaveStatusBadge } from '@/components/leave/leave-status-badge';
-import { StatCard } from '@/components/leave/stat-card';
-import { TableSkeleton } from '@/components/leave/skeletons';
-import { EmptyState } from '@/components/leave/empty-state';
+import { LeaveStatusBadge } from '@/features/leave/components/leave-status-badge';
+import { StatCard } from '@/features/leave/components/stat-card';
+import { TableSkeleton } from '@/features/leave/components/skeletons';
+import { EmptyState } from '@/features/leave/components/empty-state';
 import { useEmployeeRequests } from '@/hooks/leave/use-leave';
 import { LeaveStatus } from '@/types/leave';
 import { useCurrentUserEmployee } from '@/hooks/employee';
