+++ features/leave/components/confirmation-dialog.tsx
