+++ features/leave/components/empty-state.tsx
