+++ features/leave/components/stat-card.tsx
